Updated providers renaming database -> databaseCluster and databaseSchema -> database
