+++ src/@providers/digitalocean/modules/database.ts
@@ -2,19 +2,28 @@ import { Construct } from 'constructs';
 import { ResourceOutputs } from '../../../@resources/index.ts';
 import { ResourceModule, ResourceModuleOptions } from '../../module.ts';
 import { DataDigitaloceanDatabaseCa } from '../.gen/providers/digitalocean/data-digitalocean-database-ca/index.ts';
-import { DataDigitaloceanVpc } from '../.gen/providers/digitalocean/data-digitalocean-vpc/index.ts';
-import { DatabaseCluster } from '../.gen/providers/digitalocean/database-cluster/index.ts';
+import { DataDigitaloceanDatabaseCluster } from '../.gen/providers/digitalocean/data-digitalocean-database-cluster/index.ts';
+import { DatabaseDb } from '../.gen/providers/digitalocean/database-db/index.ts';
 import { DigitaloceanCredentials } from '../credentials.ts';
 
 export class DigitaloceanDatabaseModule extends ResourceModule<'database', DigitaloceanCredentials> {
-  database: DatabaseCluster;
   outputs: ResourceOutputs['database'];
+  db: DatabaseDb;
 
   constructor(scope: Construct, options: ResourceModuleOptions<'database', DigitaloceanCredentials>) {
     super(scope, options);
 
-    const vpc = new DataDigitaloceanVpc(this, 'vpc', {
-      id: this.inputs?.vpc || 'unknown',
+    const instance = new DataDigitaloceanDatabaseCluster(this, 'instance', {
+      name: this.inputs?.database || 'unknown',
+    });
+
+    this.db = new DatabaseDb(this, 'databaseCluster', {
+      clusterId: instance.id,
+      name: this.inputs?.name.replace(/\//g, '--') || 'unknown',
+    });
+
+    const ca = new DataDigitaloceanDatabaseCa(this, 'ca', {
+      clusterId: instance.id,
     });
 
     let protocol = this.inputs?.databaseType || 'unknown';
@@ -22,45 +31,29 @@ export class DigitaloceanDatabaseModule extends ResourceModule<'database', Digit
       protocol = 'postgresql';
     }
 
-    let engine = this.inputs?.databaseType || 'unknown';
-    if (engine === 'postgres') {
-      engine = 'pg';
-    }
-
-    this.database = new DatabaseCluster(this, 'database', {
-      name: this.inputs?.name.replaceAll('/', '--').toLowerCase() || 'unknown',
-      region: vpc.region || this.inputs?.region || 'unknown',
-      size: this.inputs?.databaseSize || 'db-s-1vcpu-1gb',
-      engine: engine,
-      version: this.inputs?.databaseVersion || '8',
-      nodeCount: 1,
-      privateNetworkUuid: vpc.id,
-    });
-
-    const ca = new DataDigitaloceanDatabaseCa(this, 'ca', {
-      clusterId: this.database.id,
-    });
-
     this.outputs = {
-      id: this.database.name,
+      id: `${instance.name}/${this.db.name}`,
+      name: this.db.name,
+      host: instance.host,
+      port: instance.port,
       protocol: protocol,
-      host: this.database.host,
-      port: this.database.port,
-      username: this.database.user,
-      password: this.database.password,
+      url: `${protocol}://${instance.host}:${instance.port}/${this.db.name}`,
+      username: instance.user,
+      password: instance.password,
       certificate: ca.certificate,
+      account: this.accountName,
     };
   }
 
   genImports(resourceId: string): Promise<Record<string, string>> {
     return Promise.resolve({
-      [this.getResourceRef(this.database)]: resourceId,
+      [this.getResourceRef(this.db)]: resourceId,
     });
   }
 
   getDisplayNames(): Record<string, string> {
     return {
-      [this.getResourceRef(this.database)]: 'Database',
+      [this.getResourceRef(this.db)]: 'Database',
     };
   }
 }
