+++ src/@providers/digitalocean/provider.ts
@@ -1,7 +1,7 @@
 import { Provider } from '../provider.ts';
 import { CldctlTestResource } from '../tests.ts';
 import { DigitaloceanCredentials, DigitaloceanCredentialsSchema } from './credentials.ts';
-import { DigitaloceanDatabaseSchemaService } from './services/database-schema.ts';
+import { DigitaloceanDatabaseClusterService } from './services/database-cluster.ts';
 import { DigitaloceanDatabaseSizeService } from './services/database-size.ts';
 import { DigitaloceanDatabaseTypeService } from './services/database-type.ts';
 import { DigitaloceanDatabaseUserService } from './services/database-user.ts';
@@ -32,9 +32,9 @@ export default class DigitaloceanProvider extends Provider<DigitaloceanCredentia
     kubernetesVersion: new DigitaloceanKubernetesVersionService(this.name, this.credentials, this.providerStore),
     kubernetesCluster: new DigitaloceanKubernetesClusterService(this.name, this.credentials, this.providerStore),
     database: new DigitaloceanDatabaseService(this.name, this.credentials, this.providerStore),
+    databaseCluster: new DigitaloceanDatabaseClusterService(this.name, this.credentials, this.providerStore),
     databaseSize: new DigitaloceanDatabaseSizeService(this.name, this.credentials, this.providerStore),
     databaseType: new DigitaloceanDatabaseTypeService(this.name, this.credentials, this.providerStore),
-    databaseSchema: new DigitaloceanDatabaseSchemaService(this.name, this.credentials, this.providerStore),
     databaseUser: new DigitaloceanDatabaseUserService(this.name, this.credentials, this.providerStore),
     databaseVersion: new DigitaloceanDatabaseVersionService(this.name, this.credentials, this.providerStore),
     dnsZone: new DigitaloceanDnsZoneService(this.name, this.credentials, this.providerStore),
