+++ src/@providers/digitalocean/services/database.ts
@@ -1,7 +1,6 @@
 import { Construct } from 'constructs';
-import { ResourceOutputs } from '../../../@resources/index.ts';
+import { ResourceOutputs } from '../../../@resources/types.ts';
 import { PagingOptions, PagingResponse } from '../../../utils/paging.ts';
-import { ResourcePresets } from '../../base.service.ts';
 import { ProviderStore } from '../../store.ts';
 import { TerraformResourceService } from '../../terraform.service.ts';
 import { DigitaloceanProvider as TerraformDigitaloceanProvider } from '../.gen/providers/digitalocean/provider/index.ts';
@@ -9,23 +8,15 @@ import { DigitaloceanCredentials } from '../credentials.ts';
 import { DigitaloceanDatabaseModule } from '../modules/database.ts';
 import { digitalOceanApiRequest } from '../utils.ts';
 
-export class DigitaloceanDatabaseService extends TerraformResourceService<'database', DigitaloceanCredentials> {
+export class DigitaloceanDatabaseService extends TerraformResourceService<
+  'database',
+  DigitaloceanCredentials
+> {
   readonly terraform_version = '1.4.5';
   readonly construct = DigitaloceanDatabaseModule;
 
   constructor(accountName: string, credentials: DigitaloceanCredentials, providerStore: ProviderStore) {
     super(accountName, credentials, providerStore);
-  }
-
-  private normalizeDatabase(database: any): ResourceOutputs['database'] {
-    return {
-      id: database.id,
-      host: database.connection.host,
-      port: database.connection.port,
-      username: database.connection.user,
-      password: database.connection.password,
-      protocol: database.engine === 'pg' ? 'postgresql' : database.engine,
-    };
   }
 
   public configureTerraformProviders(scope: Construct): TerraformDigitaloceanProvider {
@@ -35,38 +26,78 @@ export class DigitaloceanDatabaseService extends TerraformResourceService<'datab
   }
 
   async get(id: string): Promise<ResourceOutputs['database'] | undefined> {
+    if (!id.includes('/')) {
+      throw new Error(`ID must be of the format: <database>/<schema>. Got: ${id}`);
+    }
+
+    const [cluster_id, db] = id.split('/');
     const cluster = await digitalOceanApiRequest({
       credentials: this.credentials,
-      path: `/databases/${id}`,
+      path: `/databases/${cluster_id}`,
     });
-    return this.normalizeDatabase(cluster.database);
-  }
-
-  // TODO: implement filter
-  async list(
-    _filterOptions?: Partial<ResourceOutputs['database']>,
-    _pagingOptions?: Partial<PagingOptions>,
-  ): Promise<PagingResponse<ResourceOutputs['database']>> {
-    const databases = (await digitalOceanApiRequest({
+    const res = await digitalOceanApiRequest({
       credentials: this.credentials,
-      path: `/databases`,
-    })).databases;
+      path: `/databases/${cluster_id}/dbs/${db}`,
+    });
+    const protocol = cluster.database.engine === 'pg' ? 'postgresql' : cluster.database.engine;
+    const username = cluster.database.connection.user;
+    const password = '';
+    const host = cluster.database.connection.host;
+    const port = cluster.database.connection.port;
+
     return {
-      total: databases.length,
-      rows: databases.map((element: any) => this.normalizeDatabase(element)),
+      id: `${cluster_id}/${res.db.name}`,
+      protocol,
+      username,
+      password,
+      host,
+      port,
+      account: this.accountName,
+      name: res.db.name,
+      url: `${protocol}://${username}:${password}@${host}:${port}/${res.db.name}`,
     };
   }
 
-  get presets(): ResourcePresets<'database'> {
-    return [
-      {
-        display: 'Development',
-        values: {
-          databaseType: 'pg',
-          databaseVersion: '14',
-          databaseSize: 'db-s-1vcpu-1gb',
-        },
-      },
-    ];
+  async list(
+    filterOptions?: Partial<ResourceOutputs['database']>,
+    pagingOptions?: Partial<PagingOptions>,
+  ): Promise<PagingResponse<ResourceOutputs['database']>> {
+    const clusters = await digitalOceanApiRequest({
+      credentials: this.credentials,
+      path: '/databases',
+    });
+    const schemas: ResourceOutputs['database'][] = [];
+    await Promise.all(
+      (clusters.databases || []).map(async (cluster: any) => {
+        const res = await digitalOceanApiRequest({
+          credentials: this.credentials,
+          path: `/databases/${cluster.id}/dbs`,
+        });
+        res.dbs.forEach((db: any) => {
+          const protocol = cluster.engine === 'pg' ? 'postgresql' : cluster.engine;
+          const host = cluster.connection.host;
+          const port = cluster.connection.port;
+          const username = cluster.connection.user;
+          const password = '';
+
+          schemas.push({
+            id: `${cluster.id}/${db.name}`,
+            account: this.accountName,
+            host,
+            port,
+            username,
+            password,
+            protocol,
+            name: db.name,
+            url: `${protocol}://${username}:${password}@${host}:${port}/${db.name}`,
+          });
+        });
+      }),
+    );
+
+    return {
+      total: schemas.length,
+      rows: schemas,
+    };
   }
 }
