+++ src/components/v2/index.ts
@@ -292,7 +292,7 @@ export default class ComponentV2 extends Component {
         component: context.component.name,
         environment: context.environment,
         inputs: {
-          type: 'databaseSchema',
+          type: 'database',
           name: CloudNode.genResourceId({
             name: database_key,
             component: context.component.name,
